requirements.txt
blinker==1.8.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
DateTime==4.9
Flask==3.0.3
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.1
pandas==1.5.2
psycopg2-binary==2.9.5
pydataxm==0.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
tabulate==0.9.0
urllib3==1.26.13
Werkzeug==3.0.4
zope.interface==5.5.2
